+++ src/pytgpt/gpt4free/__init__.py
@@ -1,0 +1,18 @@
+from .main import GPT4FREE
+
+from .main import working_providers
+
+from .main import default_models
+
+from .main import completion_allowed_models
+
+__info__ = (
+    "Interact with various free models bundled by https://github.com/xtekky/gpt4free"
+)
+
+all = [
+    "GPT4FREE",
+    "working_providers",
+    "default_models",
+    "completion_allowed_models",
+]