+++ tests/test_gpt4free.py
@@ -1,0 +1,13 @@
+import unittest
+import tests.base as base
+from pytgpt.gpt4free import GPT4FREE
+
+
+class TestLeo(base.llmBase):
+    def setUp(self):
+        self.bot = GPT4FREE()
+        self.prompt = base.prompt
+
+
+if __name__ == "__main__":
+    unittest.main()